Loop on our reads while errno is EINTR or EAGAIN. Also make sure to
2002-06-27  Jeffrey Stedfast  <fejj@ximian.com>

	* gmime-gpg-context.c (gpg_ctx_op_step): Loop on our reads while
	errno is EINTR or EAGAIN. Also make sure to completely read stdout
	and stderr by keeping better state.
	(gpg_ctx_parse_status): If we encounted a NODATA error from gpg
	and we have diagnostics output, use that as the exception string.
	(gpg_ctx_op_exited): New function to determine if the gpg process
	has exited.
	(gpg_ctx_op_wait): Reuse any exit status information from
	gpg_ctx_op_exited().
	(gpg_verify): Check that the gpg process has not exited each loop.

+++ gmime-gpg-context.c
@@ -241,6 +241,9 @@ struct _GpgCtx {
 	
 	GByteArray *diagnostics;
 	
+	int exit_status;
+	
+	unsigned int exited:1;
 	unsigned int complete:1;
 	unsigned int seen_eof1:1;
 	unsigned int seen_eof2:1;
@@ -254,7 +257,7 @@ struct _GpgCtx {
 	unsigned int validsig:1;
 	unsigned int trust:3;
 	
-	unsigned int padding:19;
+	unsigned int padding:18;
 };
 
 static struct _GpgCtx *
@@ -271,6 +274,8 @@ gpg_ctx_new (GMimeSession *session, const char *path)
 	gpg->seen_eof1 = FALSE;
 	gpg->seen_eof2 = FALSE;
 	gpg->pid = (pid_t) -1;
+	gpg->exit_status = 0;
+	gpg->exited = FALSE;
 	
 	gpg->path = g_strdup (path);
 	gpg->userid = NULL;
@@ -772,7 +777,12 @@ gpg_ctx_parse_status (struct _GpgCtx *gpg, GMimeException *ex)
 		return -1;
 	} else if (!strncmp (status, "NODATA", 6)) {
 		/* this is an error */
-		g_mime_exception_set (ex, GMIME_EXCEPTION_SYSTEM, _("No data provided"));
+		if (gpg->diagnostics->len)
+			g_mime_exception_setv (ex, GMIME_EXCEPTION_SYSTEM, "%.*s",
+					       gpg->diagnostics->len, gpg->diagnostics->data);
+		else
+			g_mime_exception_set (ex, GMIME_EXCEPTION_SYSTEM, _("No data provided"));
+		
 		return -1;
 	} else {
 		switch (gpg->mode) {
@@ -1085,16 +1095,35 @@ gpg_ctx_op_complete (struct _GpgCtx *gpg)
 	return gpg->complete && gpg->seen_eof1 && gpg->seen_eof2;
 }
 
+static gboolean
+gpg_ctx_op_exited (struct _GpgCtx *gpg)
+{
+	pid_t retval;
+	int status;
+	
+	retval = waitpid (gpg->pid, &status, WNOHANG);
+	if (retval == gpg->pid) {
+		gpg->exit_status = status;
+		gpg->exited = TRUE;
+		return TRUE;
+	}
+	
+	return FALSE;
+}
+
 static void
 gpg_ctx_op_cancel (struct _GpgCtx *gpg)
 {
 	pid_t retval;
 	int status;
 	
+	if (gpg->exited)
+		return;
+	
 	kill (gpg->pid, SIGTERM);
 	sleep (1);
 	retval = waitpid (gpg->pid, &status, WNOHANG);
-	if (retval == 0) {
+	if (retval == (pid_t) 0) {
 		/* no more mr nice guy... */
 		kill (gpg->pid, SIGKILL);
 		sleep (1);
@@ -1109,25 +1138,30 @@ gpg_ctx_op_wait (struct _GpgCtx *gpg)
 	pid_t retval;
 	int status;
 	
-	sigemptyset (&mask);
-	sigaddset (&mask, SIGALRM);
-	sigprocmask (SIG_BLOCK, &mask, &omask);
-	alarm (1);
-	retval = waitpid (gpg->pid, &status, 0);
-	alarm (0);
-	sigprocmask (SIG_SETMASK, &omask, NULL);
-	
-	if (retval == (pid_t) -1 && errno == EINTR) {
-		/* The child is hanging: send a friendly reminder. */
-		kill (gpg->pid, SIGTERM);
-		sleep (1);
-		retval = waitpid (gpg->pid, &status, WNOHANG);
-		if (retval == (pid_t) 0) {
-			/* Still hanging; use brute force. */
-			kill (gpg->pid, SIGKILL);
+	if (!gpg->exited) {
+		sigemptyset (&mask);
+		sigaddset (&mask, SIGALRM);
+		sigprocmask (SIG_BLOCK, &mask, &omask);
+		alarm (1);
+		retval = waitpid (gpg->pid, &status, 0);
+		alarm (0);
+		sigprocmask (SIG_SETMASK, &omask, NULL);
+		
+		if (retval == (pid_t) -1 && errno == EINTR) {
+			/* The child is hanging: send a friendly reminder. */
+			kill (gpg->pid, SIGTERM);
 			sleep (1);
 			retval = waitpid (gpg->pid, &status, WNOHANG);
+			if (retval == (pid_t) 0) {
+				/* Still hanging; use brute force. */
+				kill (gpg->pid, SIGKILL);
+				sleep (1);
+				retval = waitpid (gpg->pid, &status, WNOHANG);
+			}
 		}
+	} else {
+		status = gpg->exit_status;
+		retval = gpg->pid;
 	}
 	
 	if (retval != (pid_t) -1 && WIFEXITED (status))
@@ -1135,7 +1169,6 @@ gpg_ctx_op_wait (struct _GpgCtx *gpg)
 	else
 		return -1;
 }
-
 
 
 static int
@@ -1258,7 +1291,7 @@ gpg_verify (GMimeCipherContext *context, GMimeCipherHash hash,
 		goto exception;
 	}
 	
-	while (!gpg_ctx_op_complete (gpg)) {
+	while (!gpg_ctx_op_complete (gpg) && !gpg_ctx_op_exited (gpg)) {
 		if (gpg_ctx_op_step (gpg, ex) == -1) {
 			gpg_ctx_op_cancel (gpg);
 			goto exception;
